Move errorHandlers file location
